Weapon updates
New updates to the weapons including the new katana

+++ Willam/Assets/C#/health/AllHealth.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class AllHealth : MonoBehaviour
 {
-    public float maxHealth = 100;
+    public int maxHealth = 100;
     float currentHealth;
     
     void Start()
@@ -14,7 +14,7 @@ public class AllHealth : MonoBehaviour
     }
 
 
-    public void TakeDamage(float damage)
+    public void TakeDamage(int damage)
     {
         currentHealth -= damage;
         if(currentHealth <= 0)
